(svn r5973) -Codechange: md5_append only uses size_t as nbytes param, so use that (michi_cc)

+++ md5.h
@@ -81,7 +81,7 @@ extern "C"
 void md5_init(md5_state_t *pms);
 
 /* Append a string to the message. */
-void md5_append(md5_state_t *pms, const void *data, int nbytes);
+void md5_append(md5_state_t *pms, const void *data, size_t nbytes);
 
 /* Finish the message and return the digest. */
 void md5_finish(md5_state_t *pms, md5_byte_t digest[16]);
